+++ v2/examples/lists/list_subscribe.go
@@ -1,0 +1,13 @@
+package examples
+
+import (
+	"context"
+
+	"github.com/trycourier/courier-go/v2"
+)
+
+func listSubscribe() {
+	client := courier.CreateClient("<YOUR_AUTH_TOKEN>", nil)
+
+	client.ListSubscribe(context.Background(), "my-list", "my-recipient", courier.ListSubscriptionRecipientBody{})
+}